Extract LocalRoutineRepository and associated LocalModels to the production Routine Repository
