+++ RoutineRepository/RoutineRepository/RoutineRepositoryCache/LocalExerciseRecord.swift
@@ -1,0 +1,16 @@
+//
+//  LocalExerciseRecord.swift
+//  RoutineRepository
+//
+//  Created by Boyce Estes on 10/10/22.
+//
+
+import Foundation
+
+public struct LocalExerciseRecord: Equatable {
+    
+    let id: UUID
+    
+    // relationships
+    let setRecords: [LocalSetRecord]
+}
