+++ RoutineRepository/RoutineRepository/RoutineRepositoryCache/LocalRoutine.swift
@@ -1,0 +1,19 @@
+//
+//  LocalRoutine.swift
+//  RoutineRepository
+//
+//  Created by Boyce Estes on 10/10/22.
+//
+
+import Foundation
+
+public struct LocalRoutine: Equatable {
+    
+    let id: UUID
+    let name: String
+    let creationDate: Date
+    
+    // relationships
+    let exercises: [LocalExercise]
+    let routineRecords: [LocalRoutineRecord]
+}
