+++ RoutineRepository/RoutineRepository/RoutineRepositoryCache/LocalSetRecord.swift
@@ -1,0 +1,17 @@
+//
+//  LocalSetRecord.swift
+//  RoutineRepository
+//
+//  Created by Boyce Estes on 10/10/22.
+//
+
+import Foundation
+
+public struct LocalSetRecord: Equatable {
+    
+    let id: UUID
+    let duration: Int?
+    let repCount: Int?
+    let weight: Int
+    let difficulty: Int
+}
