+++ RoutineRepository/RoutineRepository/RoutineRepositoryCache/LocalTag.swift
@@ -1,0 +1,14 @@
+//
+//  LocalTag.swift
+//  RoutineRepository
+//
+//  Created by Boyce Estes on 10/10/22.
+//
+
+import Foundation
+
+public struct LocalTag: Equatable {
+    
+    let id: UUID
+    let name: String
+}
